refactor routes into controllers and services
