+++ apps/server/src/controllers/health.js
@@ -1,0 +1,3 @@
+export function getHealth(_req, res) {
+  res.json({ ok: true });
+}
